Add BlockMap type. Break apart some Network methods into separate files.
